+++ src/Block/UrlListProvider/AbstractUrlListProvider.php
@@ -10,10 +10,18 @@ use Sobak\Scrawler\Client\Response\Elements\Url;
 abstract class AbstractUrlListProvider implements UrlListProviderInterface
 {
     /** @var Url */
+    protected $baseUrl;
+
+    /** @var Url */
     protected $currentUrl;
 
     /** @var Response */
     protected $response;
+
+    public function setBaseUrl(Url $url): void
+    {
+        $this->baseUrl = $url;
+    }
 
     public function setCurrentUrl(Url $url): void
     {