+++ tests/Functional/Crawling/ArgumentAdvancerUrlListProviderTest.php
@@ -70,6 +70,33 @@ class ArgumentAdvancerUrlListProviderTest extends ServerBasedTest
         $this->assertEquals('Fifth message', InMemoryResultWriter::$results[4]['match']);
     }
 
+    public function testWithRelativeUrl(): void
+    {
+        $host = ServerBasedTest::getHostUrl();
+
+        $config = (new Configuration())
+            ->setOperationName('test')
+            ->setBaseUrl($host)
+            ->addUrlListProvider(new ArgumentAdvancerUrlListProvider("/page-%u.html", 1, 1))
+            ->addObjectDefinition('message', new CssSelectorListMatcher('div.message'), function (ObjectConfiguration $object) {
+                $object
+                    ->addFieldDefinition('match', new CssSelectorTextMatcher('span.content'))
+                    ->addEntityMapping(SimpleMatchEntity::class)
+                    ->addResultWriter(SimpleMatchEntity::class, new InMemoryResultWriter())
+                ;
+            })
+        ;
+
+        $scrawler = new Scrawler($config, __DIR__ . '/output');
+        $scrawler->run();
+
+        $this->assertEquals('First message', InMemoryResultWriter::$results[0]['match']);
+        $this->assertEquals('Second message', InMemoryResultWriter::$results[1]['match']);
+        $this->assertEquals('Third message', InMemoryResultWriter::$results[2]['match']);
+        $this->assertEquals('Fourth message', InMemoryResultWriter::$results[3]['match']);
+        $this->assertEquals('Fifth message', InMemoryResultWriter::$results[4]['match']);
+    }
+
     public function testWithNonDefaultStep(): void
     {
         $host = ServerBasedTest::getHostUrl();